refactor: move DevWorkspace low-level logic to devworkspace package
